Updates to prepare for public release.

+++ scripts/cmake.sh
@@ -173,9 +173,12 @@ function main
   local GTEST_DIR=""
 
   if [ $TESTING = ON ] ; then
-    ln -s ../genomics_gpu/tpls/googletest-release-1.7.0.tar.gz
-    # wget -O googletest-release-1.7.0.tar.gz \
-    #   https://github.com/google/googletest/archive/release-1.7.0.tar.gz
+   if [ -e ../genomics_gpu/tpls/googletest-release-1.7.0.tar.gz ] ; then
+      ln -s ../genomics_gpu/tpls/googletest-release-1.7.0.tar.gz
+    else
+      wget -O googletest-release-1.7.0.tar.gz \
+        https://github.com/google/googletest/archive/release-1.7.0.tar.gz
+    fi
     gunzip <googletest-release-1.7.0.tar.gz | tar xf -
     GTEST_DIR=$BUILD_DIR/googletest-release-1.7.0
     mkdir $GTEST_DIR/lib
@@ -203,13 +206,19 @@ function main
 
   if [ $NOCUDA = OFF ] ; then
     local MAGMA_DIR=$BUILD_DIR/magma_patch
+    local MAGMA_VERSION=1.6.2
     if [ ! -e $MAGMA_DIR/copy_is_complete ] ; then
       rm -rf $MAGMA_DIR
       echo "Copying magma ..."
       cp -r $REPO_DIR/magma_patch $MAGMA_DIR
       # copy MAGMA source since link will be broken.
-      rm $MAGMA_DIR//magma-*.tar.gz
-      cp $REPO_DIR/tpls/magma-*.tar.gz $MAGMA_DIR/
+      rm $MAGMA_DIR//magma-${MAGMA_VERSION}.tar.gz
+      if [ -e $REPO_DIR/tpls/magma-${MAGMA_VERSION}.tar.gz ] ; then
+        cp $REPO_DIR/tpls/magma-${MAGMA_VERSION}.tar.gz $MAGMA_DIR/
+      else
+        wget -O $MAGMA_DIR/magma-${MAGMA_VERSION}.tar.gz \
+          http://icl.utk.edu/projectsfiles/magma/downloads/magma-${MAGMA_VERSION}.tar.gz
+      fi
       pushd $MAGMA_DIR
       ./create_modified_magmas.sh
       popd
@@ -324,7 +333,7 @@ function main
     C_CXX_FLAGS="$C_CXX_FLAGS -DHAVE_INT128"
   fi
   if [ $IS_IBM_AC922 = YES ] ; then
-    local TEST_COMMAND="module load $CUDA_MODULE ; env OMP_NUM_THREADS=16 jsrun -n 2 -r 1 -c 32 -g 6 -a 32 -X 1"
+    local TEST_COMMAND="module load $CUDA_MODULE ; env OMP_NUM_THREADS=1 jsrun -n 2 -r 1 -c 32 -g 6 -a 32 -X 1"
     #C_CXX_FLAGS="$C_CXX_FLAGS -DUSE_TC"
     C_CXX_FLAGS="$C_CXX_FLAGS -DHAVE_INT128"
   fi
